requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
Django==3.1.2
django-bootstrap4==2.3.1
django-crispy-forms==1.9.2
django-model-utils==4.0.0
Pillow==8.0.1
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.4.1
